feat(tg): use serde for default config
